+++ backend/src/main/java/com/example/springTrain/controller/ReviewController.java
@@ -2,7 +2,10 @@ package com.example.springTrain.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,16 +16,21 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.springTrain.dto.ReviewDto;
+import com.example.springTrain.model.User;
 import com.example.springTrain.service.ReviewService;
+import com.example.springTrain.service.UserService;
 
 @RestController
 @RequestMapping("/api/reviews")
 public class ReviewController {
 
     private final ReviewService reviewService;
+    private final UserService userService;
 
-    public ReviewController(ReviewService reviewService) {
+    public ReviewController(ReviewService reviewService,
+    		 UserService userService) {
         this.reviewService = reviewService;
+        this.userService = userService;
     }
     @GetMapping
     public ResponseEntity<List<ReviewDto>> getAllReviews() {
@@ -42,8 +50,16 @@ public class ReviewController {
     	return ResponseEntity.ok(reviewList);
     }
     
+    
     @PostMapping
     public ResponseEntity<ReviewDto> createReview(@RequestBody ReviewDto review) {
+    	Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	    String email = authentication.getName(); // Get the username (email) from the token
+	    User user = userService.findByUserEmail(email);
+	    if (user == null) {
+	        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+	    }
+	    
         ReviewDto createdReview = reviewService.createReview(review);
         return ResponseEntity.ok(createdReview);
     }
@@ -51,12 +67,26 @@ public class ReviewController {
     @PutMapping("/{id}")
     public ResponseEntity<ReviewDto> updateReview(@PathVariable("id") Long id,
     		@RequestBody ReviewDto review) {
+    	Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	    String email = authentication.getName(); // Get the username (email) from the token
+	    User user = userService.findByUserEmail(email);
+	    if (user == null) {
+	        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+	    }
+	    
     	ReviewDto updatedReview = reviewService.updateReview(id, review);
         return ResponseEntity.ok(updatedReview);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteReview(@PathVariable("id") Long id) {
+    	Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+	    String email = authentication.getName(); // Get the username (email) from the token
+	    User user = userService.findByUserEmail(email);
+	    if (user == null) {
+	        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+	    }
+	    
         reviewService.deleteReview(id);
         return ResponseEntity.noContent().build();
     }
